Update git data in filenames

+++ pdmpk/args.h
@@ -28,10 +28,10 @@ struct Args {
   int nlevel;
 
   /// Index to select the mirror method.
-  int mirror_method;
+  size_t mirror_method;
 
   /// Index to select the weight update method.
-  int weight_update_method;
+  size_t weight_update_method;
 
   /// Metis options.
   idx_t opt[METIS_NOPTIONS];
